add invoices

+++ src/EgyptianeInvoicing.Shared/Dtos/ImportedInvoiceItemDto.cs
@@ -13,10 +13,10 @@
         public string Currency { get; set; } = "جنيه مصري"; // العملة
         public double? CurrencyConvert { get; set; } // سعر الوحدة بالعملة الاجنبية
         public string? VATCode { get; set; }                // كود ضريبة القيمة المضافة
-        public decimal VATPercentage { get; set; } = 0;         // نسبة ضريبة القيمة المضافة
-        public decimal RelativeTableTaxPercentage { get; set; } = 0; // نسبة ضريبة الجدول النسبية
-        public decimal SpecificTableTaxPercentage { get; set; } = 0; // نسبة ضريبة الجدول النوعية
+        public double VATPercentage { get; set; } = 0;         // نسبة ضريبة القيمة المضافة
+        public double RelativeTableTaxPercentage { get; set; } = 0; // نسبة ضريبة الجدول النسبية
+        public double SpecificTableTaxPercentage { get; set; } = 0; // نسبة ضريبة الجدول النوعية
         public string? DiscountTaxCode { get; set; }        // كود الخصم تحت حساب الضريبة
-        public decimal DiscountTaxPercentage { get; set; } = 0; // نسبة الخصم تحت حساب الضريبة
+        public double DiscountTaxPercentage { get; set; } = 0; // نسبة الخصم تحت حساب الضريبة
     }
 }
